[ui] Move custom React hooks into UI directory
