+++ src/batch/monitor.tsx
@@ -23,13 +23,13 @@ import { TaskSelectorClient } from 'batch/client/task_selector';
 import { MoneyTracker, primedMoneyTracker } from 'util/money-tracker';
 
 import { extend } from 'util/extend';
-import { useNsUpdate, usePoll, useTheme } from 'util/hooks';
 import { BaseServer, type Handlers } from 'util/protocol';
 import { HUD_HEIGHT, HUD_WIDTH, STATUS_WINDOW_WIDTH } from 'util/ui';
 
-import { React } from 'lib/react';
-
 import { UI_CONFIG } from 'ui/config';
+import { useNsUpdate, usePoll, useTheme } from 'ui/hooks';
+
+import { React } from 'lib/react';
 
 const FLAGS = [
     ['refreshrate', 200],
